+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [unreleased]
 
+## [1.0.1]
+
 - Fix parsing of inline code in Bison grammars.
 
 - Fix spurious "duplicate description" warning in multithreaded builds.
@@ -14,6 +16,7 @@
 
 - Initial release.
 
-[unreleased]: https://github.com/taminomara/sphinx-syntax/compare/v1.0.0-post1...HEAD
+[unreleased]: https://github.com/taminomara/sphinx-syntax/compare/v1.0.1...HEAD
+[unreleased]: https://github.com/taminomara/sphinx-syntax/compare/v1.0.0-post1...v1.0.1
 [1.0.0-post1]: https://github.com/taminomara/sphinx-syntax/compare/v1.0.0...v1.0.0-post1
 [1.0.0]: https://github.com/taminomara/sphinx-syntax/releases/tag/v1.0.0